Add hero ship and parent game object to rotate to proper orientation
Select the heroShip child of _Hero and set its transform to P:[ 0, 0, 0 ] R:[ -90, 0, 0 ] S:[ 0.01, 0.01, 0.01 ].

This rotates the heroShip model so that it’s oriented correctly for our game. You probably noticed the scale of 0.01 in the heroShip Transform. That is because the main unit in many modeling programs is centimeters, whereas in Unity, it is meters. So, we need to reduce the size of the heroShip model by 100x (i.e., a scale of 0.01) to make it the size we want. Now, the model is oriented properly, and it’s time to add more components to the enclosing _Hero GameObject.